Rename "io" constants and stl subpackage to "ui"
